% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Bladder cancer microarray data}
\format{
microarray \code{data_matrix},whose row contains 54675 features and column contains 12 samples, where 3 normal groups and 9 tumor groups
}
\usage{
bladder
}
\description{
Bladder cancer microarray data
}
\examples{
data(bladder)
}
\references{
L. Mengual, M. Burset, E. Ars, J.J. Lozano, H. Villavicencio, M.J. Ribal, A. Alcaraz, DNA Microarray Expression Profiling of Bladder Cancer Allows Identification of Noninvasive Diagnostic Markers, Journal of Urology 182 (2009) 741–748. https://doi.org/10.1016/j.juro.2009.03.084.
}
\keyword{datasets}
