% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GS.R
\name{GS}
\alias{GS}
\title{Identify Key features}
\usage{
GS(
  Data,
  n_groups,
  DE_ANOVA = TRUE,
  DE_KW = TRUE,
  DE_Ttest = TRUE,
  DE_Sam = TRUE,
  DE_limma = TRUE,
  P = 0.05,
  Adjp = "none",
  N_top = 15
)
}
\arguments{
\item{Data}{Microarray data matrix, whose row contain features/genes and column contain samples}

\item{n_groups}{Number of groups eg. n1=n2=3; n_groups <- c(3,3)}

\item{DE_ANOVA}{Analysis of variance (ANOVA) by default TRUE}

\item{DE_KW}{Kruskal–Wallis (KW) test) by default TRUE}

\item{DE_Ttest}{T-test) by default TRUE}

\item{DE_Sam}{Significance analysis of microarrays (SAM) by default TRUE}

\item{DE_limma}{Linear models for microarray data (LIMMA) by default TRUE}

\item{P}{pvalue by default=0.05}

\item{Adjp}{adjusted correction method see also: \code{\link[stats]{p.adjust}}}

\item{N_top}{Number of significant features/genes selection by default 15}
}
\value{
Top significant features/genes matrix
}
\description{
This function is used to identify key features/genes.
}
\examples{
top.df<-HMCC::GS(bladder,n_groups =c(3,9))
col <- colorRampPalette(c("green", "black", "red"))(100)
heatmap(as.matrix(t(top.df[,-16])),scale = "none",col = col,Rowv = NA,Colv=NA)

#Classification
attach(top.df)

#SVM
library(e1071)
model.svm <- svm(DataLabels ~ ., data = top.df)
summary(model.svm)
x <- subset(top.df, select = -DataLabels)
pred.svm <- predict(model.svm, x,type = "class")
table(predict = pred.svm, actual =DataLabels)

#random forest
library(randomForest)
model.rf <- randomForest(DataLabels~.,data=top.df)
pred.rf <- predict(model.rf, x,type = "class")
table(predict = pred.rf, actual =DataLabels)

}
\seealso{
\code{\link[e1071]{svm}}, \code{\link[randomForest]{randomForest}}
}
